Move type declarations into src/@types/*
