+++ src/@types/worker-loader/index.d.ts
@@ -1,0 +1,10 @@
+/** Webpack worker-loader type setup */
+declare module 'worker-loader!*' {
+  /** Worker */
+  class WebpackWorker extends Worker {
+    /** Webpack's Worker constructor */
+    public constructor();
+  }
+
+  export default WebpackWorker;
+}
